create pages folder, reorganize components
